requirements.txt
growwapi>=1.5.0
pyotp>=2.9.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=6.0.0
